Compiler warning/error fix for Chromium
git-svn-id: http://skia.googlecode.com/svn/trunk@11494 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/core/SkPathRef.h
@@ -347,11 +347,11 @@ public:
      * Gets the number of bytes that would be written in writeBuffer()
      */
     uint32_t writeSize() {
-        return 5 * sizeof(uint32_t) +
-               fVerbCnt * sizeof(uint8_t) +
-               fPointCnt * sizeof(SkPoint) +
-               fConicWeights.bytes() +
-               sizeof(SkRect);
+        return uint32_t(5 * sizeof(uint32_t) +
+                        fVerbCnt * sizeof(uint8_t) +
+                        fPointCnt * sizeof(SkPoint) +
+                        fConicWeights.bytes() +
+                        sizeof(SkRect));
     }
 
 private:
